block: sed-opal: Add code to handle keyrings for PEKs.
Create a keyring ".sed_opal" to serve as the container for
the default PEKs.

If sed_ioctl() is called with opal_key.key_type set to OPAL_KEYRING,
then the PEK will be acquired from an existing keyring. If
opal_key.key_len is zero, use the default PEKs from the ".sed_opal"
keyring. If opal_key.key_len is the size of a key serial number,
extract the serial number and lookup key directly.